requirement.txt
asgiref==3.8.1
Django==5.1.5
django-filter==24.3
django-ipware==7.0.1
django-utils-six==2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-ipware==3.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
